Renamed agentClass.ts into class.ts
This light refactoring makes the file names more distinctive.
